Make the live session more like Nixbook, and bundle all* required Nixbook packages in the ISO's nix store
There might be some missing, I'm not sure.

The installer ISO now has Cinnamon instead of GNOME.

The installer ISO (and Nixbook) now has audio support for Chromebooks that use SOF out of the box. Required packages and configuration for running MrChromebox's Firmware Utility Script and flashing firmware is also included.

On my laptop with i7-10610U, it took ~3 hr to build the ISO, and the generated iso is almost 30GB.

I tested out the installer in a VM and it downloaded about 150MB when installing Nixbook (not lite). I think this is definitely an improvement.